test: add `setup-e2e-test.ts` to avoid repeating setup codes
